+++ tests/knight-moves.json
@@ -1,0 +1,38 @@
+[
+  {
+    "description": "white knight",
+    "hexchess": "1/3/5/2P1p2/9/5N5/11/11/11/11/11 w - 0 1",
+    "from": "f6",
+    "result": [
+      "f6g8",
+      "f6h7",
+      "f6i5",
+      "f6i4",
+      "f6h3",
+      "f6g3",
+      "f6e3",
+      "f6d3",
+      "f6c4",
+      "f6c5",
+      "f6d7"
+    ]
+  },
+  {
+    "description": "black knight",
+    "hexchess": "1/3/5/2P1p2/9/5n5/11/11/11/11/11 b - 0 1",
+    "from": "f6",
+    "result": [
+      "f6h7",
+      "f6i5",
+      "f6i4",
+      "f6h3",
+      "f6g3",
+      "f6e3",
+      "f6d3",
+      "f6c4",
+      "f6c5",
+      "f6d7",
+      "f6e8"
+    ]
+  }
+]
