+++ tests/rust/mod.rs
@@ -1,4 +1,4 @@
 mod board_traversal;
 mod graph_traversal;
+mod piece_movement;
 mod positions;
-mod straight_line_moves;
